# src/main/java/com/example/bankcards/dto/AuthDto.java
package com.example.bankcards.dto;

import jakarta.validation.constraints.NotBlank;

public class AuthDto {
    
    @NotBlank(message = "Имя пользователя обязательно")
    private String username;
    
    @NotBlank(message = "Пароль обязателен")
    private String password;

    public AuthDto() {}

    public AuthDto(String username, String password) {
        this.username = username;
        this.password = password;
    }

    // Getters and Setters
    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}